refactor: move JsonConverter constructor parameter tests to a dedicated file
