+++ json_serializable/test/src/json_converter_test_input.dart
@@ -153,38 +153,6 @@ class _DurationMillisecondConverter implements JsonConverter<Duration, int> {
 }
 
 @ShouldGenerate(r'''
-JsonConverterCtorParams _$JsonConverterCtorParamsFromJson(
-  Map<String, dynamic> json,
-) => JsonConverterCtorParams()
-  ..value = const _ConverterWithCtorParams(
-    42,
-  ).fromJson((json['value'] as num).toInt());
-
-Map<String, dynamic> _$JsonConverterCtorParamsToJson(
-  JsonConverterCtorParams instance,
-) => <String, dynamic>{
-  'value': const _ConverterWithCtorParams(42).toJson(instance.value),
-};
-''')
-@JsonSerializable()
-@_ConverterWithCtorParams(42)
-class JsonConverterCtorParams {
-  late Duration value;
-}
-
-class _ConverterWithCtorParams implements JsonConverter<Duration, int> {
-  final int param;
-
-  const _ConverterWithCtorParams(this.param);
-
-  @override
-  Duration fromJson(int json) => throw UnimplementedError();
-
-  @override
-  int toJson(Duration object) => 0;
-}
-
-@ShouldGenerate(r'''
 Map<String, dynamic> _$JsonConverterOnGetterToJson(
   JsonConverterOnGetter instance,
 ) => <String, dynamic>{
@@ -230,100 +198,4 @@ class _NullableConverter implements JsonConverter<NeedsConversion?, Object?> {
 
   @override
   Object? toJson(NeedsConversion? object) => null;
-}
-
-@ShouldGenerate(r'''
-JsonConverterStringParams _$JsonConverterStringParamsFromJson(
-  Map<String, dynamic> json,
-) => JsonConverterStringParams()
-  ..value = const _ConverterWithStringParams(
-    'cool',
-  ).fromJson((json['value'] as num).toInt());
-
-Map<String, dynamic> _$JsonConverterStringParamsToJson(
-  JsonConverterStringParams instance,
-) => <String, dynamic>{
-  'value': const _ConverterWithStringParams('cool').toJson(instance.value),
-};
-''')
-@JsonSerializable()
-@_ConverterWithStringParams('cool')
-class JsonConverterStringParams {
-  late Duration value;
-}
-
-class _ConverterWithStringParams implements JsonConverter<Duration, int> {
-  final String param;
-
-  const _ConverterWithStringParams(this.param);
-
-  @override
-  Duration fromJson(int json) => throw UnimplementedError();
-
-  @override
-  int toJson(Duration object) => 0;
-}
-
-@ShouldGenerate(r'''
-JsonConverterListParams _$JsonConverterListParamsFromJson(
-  Map<String, dynamic> json,
-) => JsonConverterListParams()
-  ..value = const _ConverterWithListParams([
-    42,
-  ]).fromJson((json['value'] as num).toInt());
-
-Map<String, dynamic> _$JsonConverterListParamsToJson(
-  JsonConverterListParams instance,
-) => <String, dynamic>{
-  'value': const _ConverterWithListParams([42]).toJson(instance.value),
-};
-''')
-@JsonSerializable()
-@_ConverterWithListParams([42])
-class JsonConverterListParams {
-  late Duration value;
-}
-
-class _ConverterWithListParams implements JsonConverter<Duration, int> {
-  final List<int> param;
-
-  const _ConverterWithListParams(this.param);
-
-  @override
-  Duration fromJson(int json) => throw UnimplementedError();
-
-  @override
-  int toJson(Duration object) => 0;
-}
-
-@ShouldGenerate(r'''
-JsonConverterNamedParams _$JsonConverterNamedParamsFromJson(
-  Map<String, dynamic> json,
-) => JsonConverterNamedParams()
-  ..value = const _ConverterWithNamedParams(
-    param: 42,
-  ).fromJson((json['value'] as num).toInt());
-
-Map<String, dynamic> _$JsonConverterNamedParamsToJson(
-  JsonConverterNamedParams instance,
-) => <String, dynamic>{
-  'value': const _ConverterWithNamedParams(param: 42).toJson(instance.value),
-};
-''')
-@JsonSerializable()
-@_ConverterWithNamedParams(param: 42)
-class JsonConverterNamedParams {
-  late Duration value;
-}
-
-class _ConverterWithNamedParams implements JsonConverter<Duration, int> {
-  final int param;
-
-  const _ConverterWithNamedParams({required this.param});
-
-  @override
-  Duration fromJson(int json) => throw UnimplementedError();
-
-  @override
-  int toJson(Duration object) => 0;
 }
